rename: input 폴더 제거, 내부 컴포넌트를 분리하여 각각 AuthInputNormal과 AuthInputPassword로 폴더명 변경
