--- src/components/common/Auth/Inputs/index.tsx
@@ -1,2 +1,0 @@
-export { default as InputNormal } from './InputNormal';
-export { default as InputPassword } from './InputPassword';
