+++ src/components/common/Auth/index.tsx
@@ -1,1 +1,3 @@
 export { default as AuthButton } from './AuthButton';
+export { default as AuthInputNormal } from './AuthInputNormal';
+export { default as AuthInputPassword } from './AuthInputPassword';
